% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerui.R
\name{pageruiInput}
\alias{pageruiInput}
\title{Create a pager-ui Input control}
\usage{
pageruiInput(inputId, page_current = NULL, pages_total = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value}

\item{page_current}{The currently selected page}

\item{total_pages}{The total number of pages available}
}
\description{
Create a pager-ui Input control
}
\seealso{
\link{updatePageruiInput}
}
