% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerui.R
\name{updatePageruiInput}
\alias{updatePageruiInput}
\title{Change the value of a pager-ui input on the client}
\usage{
updatePageruiInput(session, inputId, page_current = NULL,
  pages_total = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}}

\item{inputId}{The id of the input object}

\item{page_current}{Current page value}

\item{pages_total}{Total pages value}
}
\description{
Change the value of a pager-ui input on the client
}
\seealso{
\link{pageruiInput}
}
